# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "description": "Crea un pedido OPEN/PENDING con numeración de tienda y precios actuales de productos. Descuenta inventario controlado; stock insuficiente o producto inexistente produce 400. Usa la ubicación activa predeterminada. Roles permitidos: OWNER, CASHIER, MANAGER.",
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada (POST devuelve 201).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, CASHIER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear pedido",
        "tags": [
          "orders"
        ],
        "x-roles": [
          "OWNER",
          "CASHIER",
          "MANAGER"
        ]
      },
      "get": {
        "description": "Busca por nombre del pedido y nombre, apellido, correo o teléfono del cliente. Filtra por status. startDate y endDate se validan pero actualmente NO se aplican al listado. Orden predeterminado createdAt desc. Roles permitidos: OWNER, CASHIER, MANAGER.",
        "operationId": "OrdersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página solicitada; enviar un entero positivo",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página; enviar un entero positivo",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Texto de búsqueda, sin distinguir mayúsculas",
            "schema": {
              "example": "cafe",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo del modelo Prisma para ordenar; no usar nombres calculados de la respuesta",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "description": "Dirección del orden",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Estado o rol permitido",
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Aceptado y validado, pero actualmente no se aplica al listado. Usar /orders/stats para filtrar estadísticas por fecha.",
            "schema": {
              "example": "2026-09-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Aceptado y validado, pero actualmente no se aplica al listado. Usar /orders/stats para filtrar estadísticas por fecha.",
            "schema": {
              "example": "2026-09-27",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPage"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, CASHIER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar pedidos",
        "tags": [
          "orders"
        ],
        "x-roles": [
          "OWNER",
          "CASHIER",
          "MANAGER"
        ]
      }
    },
    "/api/orders/stats": {
      "get": {
        "description": "Filtra por fecha de creación e incluye los extremos del día según la zona horaria del servidor. Sin fechas usa el día actual; con un solo extremo queda abierto el otro. orders, sales e items excluyen cancelados; cancelledOrders los cuenta aparte. sales suma totales de pedidos, no pagos cobrados. Roles permitidos: OWNER, CASHIER, MANAGER.",
        "operationId": "OrdersController_getOrderStats",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Inicio inclusivo por fecha de creación, normalizado al inicio del día del servidor",
            "schema": {
              "example": "2026-09-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Fin inclusivo por fecha de creación, normalizado al final del día del servidor",
            "schema": {
              "example": "2026-09-27",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStats"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, CASHIER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar estadísticas de pedidos",
        "tags": [
          "orders"
        ],
        "x-roles": [
          "OWNER",
          "CASHIER",
          "MANAGER"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "description": "Obtiene el registro identificado por UUID dentro de la tienda del usuario. Roles permitidos: OWNER, CASHIER, MANAGER.",
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, CASHIER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar pedido",
        "tags": [
          "orders"
        ],
        "x-roles": [
          "OWNER",
          "CASHIER",
          "MANAGER"
        ]
      },
      "patch": {
        "description": "Actualiza customerId, email, phone y note. items es aceptado y validado, pero ignorado: use las rutas de artículos. No comprueba el estado del pedido. Roles permitidos: OWNER, CASHIER, MANAGER.",
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, CASHIER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar datos del pedido",
        "tags": [
          "orders"
        ],
        "x-roles": [
          "OWNER",
          "CASHIER",
          "MANAGER"
        ]
      }
    },
    "/api/orders/{id}/items": {
      "post": {
        "description": "Requiere pedido OPEN. Agrega una línea con cantidad 1 por cada productId, descuenta inventario controlado y recalcula totales. No fusiona líneas existentes. Roles permitidos: OWNER, CASHIER, MANAGER.",
        "operationId": "OrdersController_addItems",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItemsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada (POST devuelve 201).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, CASHIER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agregar artículos",
        "tags": [
          "orders"
        ],
        "x-roles": [
          "OWNER",
          "CASHIER",
          "MANAGER"
        ]
      }
    },
    "/api/orders/{id}/items/{itemId}": {
      "patch": {
        "description": "Requiere pedido OPEN. quantity es la nueva cantidad absoluta; cero retira el artículo de la respuesta. Aumentar descuenta stock; reducir solo repone si restock=true (predeterminado false). Recalcula totales. Un itemId inexistente puede producir 500 por Prisma no interceptado. Roles permitidos: OWNER, CASHIER, MANAGER.",
        "operationId": "OrdersController_editItemQuantity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "UUID del artículo dentro del pedido.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditItemQuantityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, CASHIER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambiar cantidad de un artículo",
        "tags": [
          "orders"
        ],
        "x-roles": [
          "OWNER",
          "CASHIER",
          "MANAGER"
        ]
      }
    },
...
  },
  "info": {
    "title": "La Hermandad — API POS",
    "description": "API del punto de venta: autenticación, usuarios, categorías, productos, clientes y pedidos con pagos. Las rutas protegidas obtienen la tienda desde el usuario autenticado. Use login o register y pegue el token en Authorize. Los JWT duran 24 horas. Los cuerpos rechazan propiedades desconocidas. Los importes de entrada son números y los Decimal de salida son cadenas. Todos los POST devuelven 201; PATCH y DELETE devuelven 200 con cuerpo. Los permisos se indican mediante x-roles en cada operación.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Pegue únicamente el token obtenido en auth/login o auth/register."
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "description": "Identificador UUID del recurso",
            "example": "123e4567-e89b-42d3-a456-426614174000",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico",
            "example": "alvaro@example.com",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "description": "Teléfono en formato internacional",
            "example": "+59171234567"
          },
          "note": {
            "type": "string",
            "description": "Nota interna",
            "example": "Atender por la tarde"
          },
          "items": {
            "description": "Artículos del pedido",
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "CLOSED",
          "CANCELLED"
        ]
      },
...
      "AddItemsDto": {
        "type": "object",
        "properties": {
          "productIds": {
            "description": "Productos a agregar, cada uno con cantidad 1",
            "example": [
              "123e4567-e89b-42d3-a456-426614174000"
            ],
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "productIds"
        ]
      },
...
      "Error": {
        "type": "object",
        "required": [
          "statusCode",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "description": "Mensaje del servicio o lista de errores de validación.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "property unknown should not exist"
            ]
          },
          "error": {
            "type": "string",
            "description": "Nombre HTTP del error, cuando Nest lo incluye",
            "example": "Bad Request"
          }
        }
      },
...
      "OrderPage": {
        "type": "object",
        "required": [
          "meta",
          "results"
        ],
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        }
      }
    }
  }
}